requirements.txt
streamlit
mysql-connector-python
langchain
langchain-community
langchain-core
langchain-groq
groq